requirements.txt
Django==4.1.7
djangorestframework==3.14.0
django-environ==0.9.0
django-cors-headers~=3.13.0
psycopg2-binary==2.9.5
requests==2.28.1
Pillow==9.2.0
pytz==2022.7.1
pyJWT==2.7.0